Allow overriding pod security admission label
Introduces an annotation to override the pod security admission label for
hosted control plane namespaces.

This is a follow up to https://github.com/openshift/hypershift/pull/2778
where the default PSA is now changed to 'Restricted'. For some consumers
like IBM, this label may be too restrictive for other workloads they may
want to run in the control plane namespace.