Update Etcd health check to mirror standalone etcd
The previous etcd health check uses an exec probe. Because we don't set
a timeout, the timeout defafults to 1. This commit modifies the etcd
statefulset to use a sidecar container with the cluster etcd operator
image that implements an http get type of probe. It also modifies the
timeouts to mirror those set in the standalone case.