HOSTEDCP-1000: Add user CA(s) to ocm-trusted-ca-bundle ConfigMap (#2713)
* HOSTEDCP-1000: Add image to management cluster capabilities

https://issues.redhat.com/browse/HOSTEDCP-1000
image.config.openshift.io will need to be checked for
the CAs for the image registries a user specifies.

* HOSTEDCP-1000: Add user CA(s) to ocm-trusted-ca-bundle ConfigMap

https://issues.redhat.com/browse/HOSTEDCP-1000
Adds the CA certificates for mirror registries specified by the user
to the ocm-trusted-ca-bundle injected in
https://github.com/openshift/hypershift/pull/2437 so that these
certificates may be used to pull images from these mirror registries.

* HOSTEDCP-1000: Create Init container for hypershift operator

https://issues.redhat.com/browse/HOSTEDCP-1000
Adds an init container to the hypershift operator to initialize
the operator's environment. The init container includes changes for
adding the user's CA bundle to the ocm-trusted-ca-bundle.