Use single timeout for resource cleanup
Instead of using individual resource timeouts, simply give up after an
overall timeout of attempting to clean up resources. The HCCO no longer
updates the lastTransitionTimeout of the CloudResourcesDestroyed
condition, and the CPO uses the last time it was updated (if available)
to determine whether it should give up or not.