Drop default kas port to 443 and endpoint reconciliation for new clusters
With this change, existing HCs defaulting Networking.APIServer.Port to 443 will preserve current behaviour of overriding the kas endpoint in the guest cluster.
This is to avoid rollouts that would result from changing the haproxy config.

New HCs wont default to 443 and they won't override the kas endpoint. Seeting Networking.APIServer.Port to 443 by a consumer is still techincally possible but unrecommended.