OCPBUGS-14611: Ignored cloud dangling objects after timeout on HC teardown (#2894)
* OCPBUGS-14611: Ignored cloud dangling objects after timeout on HC teardown

This patch basically adds a HostedCluster destroy grace period when there are cloud dangling objects that hypershift cannot delete. It includes a default timeout, initially set to 0 secods, and it should be increased depending of the deployment needs.

If the Time.Now() reaches the DeletionTimestamp + timeout (of that concrete object), the controller mark that object as ignored and it will continue with deprovisioning.
Cloud objects involved:

- ImageRegistryStorage
- IngressControllers
- LoadBalancers
- PersistentVolumes

When the cluster is marked as deleted but not finalized, a new condition is created "HostedClusterDestroyed". That condition will give extra time for OCM to grab the non deleted cloud dangling objects and send a notification to the final user. After that the finalizer is deleted and the HostedCluster fully destroyed

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>

* Use single timeout for resource cleanup

Instead of using individual resource timeouts, simply give up after an
overall timeout of attempting to clean up resources. The HCCO no longer
updates the lastTransitionTimeout of the CloudResourcesDestroyed
condition, and the CPO uses the last time it was updated (if available)
to determine whether it should give up or not.

---------

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>
Co-authored-by: Cesar Wong <cewong@redhat.com>