[kubevirt platform] Add HostedCluster condition for Valid/Invalid Infra Network MTU
When using an hosted cluster based on a kubevirt provider, we recommend having a management cluster with MTU of 9000 bytes or grater in order for the hosted cluster to function properly.
If a smaller sized MTU is used, a performance degragation is expected. We should bring visibility when this occurs, in terms of a new condition on the HostedCluster resource

Signed-off-by: Oren Cohen <ocohen@redhat.com>