Add flag to create a single NAT gateway
When creating multi-zone hosted clusters, the --single-nat-gateway flag
allows creating a single nat gateway instead of 3 and configures all
cluster traffic to flow through that gateway. This is meant to be used
in development and testing so that we can avoid spending on a nat
gateway for every zone of a hosted cluster.