Use MCO and CCO image references when looking up mappings
Currently, we unconditionally use an image mapping from the management
cluster if a mapping exists for ocp-release-dev or ocp/release.
When the individual images do not use those registries, the wrong
mapping is used.
This commit uses the individual images from the payload to look for any
image mappings that may exist on the management cluster.