Dump kubevirt external infra clusters
When using kubevirt provider with an exteranl infra cluster, the dump
now will include the info for this cluster as well.

In this case, the dump directory will include a new
`external-infra-clusters` directory, containing a directory for each
external infra cluster used; e.g.

dump-dir
  +- external-infra-clusters
      +- nodepool-<node pool UID>

Also, if kubevirt is not installed on the management cluster, the dump
log won't include errors about unknown resource types.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>