OCPBUGS-18720: amend OLM catalogs ImageStream according to annotation
With https://github.com/openshift/hypershift/pull/2947
we introduced a mechanism to amend the OLM catalog
ImageStream according to OpenShiftImageRegistryOverrides
that are always automatically initializaed according to
any ImageContentSourcePolicy (ICSP) or any ImageDigestMirrorSet (IDMS)
instance from an OpenShift management cluster.

But assuming that the whole default OLM catalogs
got properly mirrored to a different registry
just because an ICSP is there on the management cluster is
overblown and with the imagestream the replacements cannot be
validated ahead ot time.

Let's not do it by default according to ICSP/IDMS
on the management cluster but only according to the
value of a new explicit annotation.

Automatically mirror the 5 involved annotations
from the HC CR to the HCP one.

Fixes: https://issues.redhat.com/browse/OCPBUGS-18720

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>