[Kubevirt] Remove EgressFirewall Creation in HCP namespace
Starting from 4.14, the EgressFirewall is not needed any longer, since kubevirt platform consumes the kubevirt provider in RHCOS, thus an attempt to access the metadata server is not being initiated, as opposed to the legacy OpenStack RHCOS provider.
Therefore, EgressFirewall is not needed to be deployed on the mgmt/infra cluster for guests >= v4.14.

Signed-off-by: Oren Cohen <ocohen@redhat.com>