CNV-34094: Add validating webhook
Add validating webhook for HostedClusters and NodePools, to check if
kubevirt external infra cluster is with sufficient CNV and K8s versions.

This is already checked on the operator, but adding this logic to the
webhook will prevent the creating of resources in insufficient
environment, in advance. Fixing this error is not an easy task and
require upgrading the infra cluster, or the CNV installation, and we
prefer to fail fast in this case.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>