KubeVirt: create the etcd encryption key secret, if missing
To allow creating of KubeVirt hosted cluster using the hosted cluster
API (rather than using the cli).

When creating the hosted cluster using the cli, the cli also creates the
secret. But when creating the hosted cluster using the hosted cluster
API, the secret is not created.

This PR changes hypershift so it now creates the etcd
encryption key secret, if it is not already exist.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>