Fix a bug in the validating webhook
When creating a HostedCluster from command line, the HostedCluster CR is
one of the first resources that been created. Some resources are created
later but are actually needed.

The new validating webhook need the secret that holds the kubeconfig of
the infra-cluster (when creating a a cluster with an external KubeVirt
infra cluster). The webhook is called for HostedCluster cration, while
the secret is not created yet, and so the creation of the HostedCluster
fails.

This PR changes the order of the resources when creatig a new cluster,
so all the dependencies are created before the HostedCluster CR.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>