Add 'hypershift.openshift.io/hcp: true' label for nodeSelectors
HOSTEDCP-1316 outlines a use-case where a Daemonset needs to run on
management clusters and be able to select request-serving nodes that
expect to have HCP pods running on them with a nodeSelector, which
requires a fixed key-value pair as a label.

The existing label, hypershift.openshift.io/cluster= ${HCP_NS}
unfortunately does not work until nodeAffinity is supported in the other
use-case. When that is true, this label is free to be removed.

Signed-off-by: Michael Shen <mshen@redhat.com>