HOSTEDCP-1273: Added support for OLM Disable default sources on HC creation
The pull request (PR) embeds the OperatorHubSpec field into the hostedcluster.Spec.Configuration API field. This allows us to disable/enable default sources in Management or Guest placement only during creation.

This functionality is also incorporated into the CLIs through a flag called .
However, it's important to note that the OperatorHub capability to disable/enable individual default sources is not covered in this PR. As a result, that field will be disregarded in the CPO and HCCO reconciliation.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>