[Release 4.14] Fix a bug on deletion of a hostedcluster
A user destroying a HostedCluster can cause the HostedCluster to hang
indefinitely if the destroy command times out during execution.

This is due to the hcp cli placing a finalizer on the HostedCluster
during deletion which the cli tool later removes after waiting for
some clean up actions to occur. If a user cancels the hcp destroy
cluster command (or the command times out) while the cli is waiting
for cleanup, then the HostedCluster will hang indefinitely with
a DeletionTimestamp != nil.

This PR only performs the additional cleaning logic, including the
setting and removing the special finalizer, for platform that actually
creates additional resources. Also, this PR sets the hostedCluster as
the owner for these secrets, so they will be deleted with the
HostedCluster.

This PR is a manual cherry-pick of #3234

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>