kubevirt: Use selector less services for ingress
The KubeVirt provider default ingress uses a passthrow service to
connect to the hosted cluster ingress using it's ingress service
nodeport, this works fine with pod network but when using secondary
networks custom endpointslices has to be created and maintained. This change
remove the selector from the service and maintain the endpointslices at a
reconcile cycle at hosted control plane config oeprator.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>