OCPBUGS-28538: Add docs for other SDN providers (Cilium & Calico)
This PR includes Cilium and Calico SDNs to have them configured in the
Hosted Cluster. In the Cilium part also covers some conformance tests
that could be deployed into the HostedCluster to validate that the SDN
is working as expected.

Signed-off-by: Mario Vazquez <mavazque@redhat.com>
Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>

This commit includes the fixes for versions 4.14.10 and also from 4.15.z, making Cilium work as SDN properly. Also included configurations for IPv6 and Dual Stack.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>