hypershift-operator: allow injecting CPO image labels
When the control-plane-operator image is injected via annotation, it is
possible to have the system use an image that resolves at the node level
and runs, but is not query-able with the normal image metadata client.
In order to not require the image to be present in some real registry,
we can furthermore allow an annoation to provide the content that
normally would be determined by resolving the `LABEL` data from image
metadata.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>