Set outbound LoadBalancer name on AzureCluster
This fixes the issue of Azure VMs not joining a HostedCluster as nodes
because they were not getting added to the outbound LoadBalancer in
their resource group. This change sets the name of the outbound
LoadBalancer in the AzureCluster so that the CAPI provider can set the
proper backend pool on VM NICs.