Always include AWS default security group in worker security groups
Before this commit, the default security group was only used if no
security groups were specified on a NodePool. After this change, the
default security group will always be used on all workers. Any security
groups defined in the NodePool will be added to the default.

Security group creation has been removed from the create infra CLI,
given that it's a duplicate of the default one created by the control
plane operator.