Remove CLI inspection of release image
Previously, the hcp client tool would dynamically
inspect the release image to determine what the
version was in use, and use that information
to default the network type.

That inspection of the release image causes
issues in the disconnected use case.
It's possible that the machine someone is
executing the client on can't directly
access the image registry.
In this situation, the client fails and
prevents a user from being able to create a cluster.

Moving forward, we need the client to only
require access to the api server of the mgmt cluster.

Signed-off-by: David Vossel <davidvossel@gmail.com>
Signed-off-by: Simone Tiraboschi <stirabos@gmail.com>