CLI: create nodepool aws - remove sg requirement
A security group is no longer created by the CLI when creating a new
cluster since we rely on the default one created by the control plane
operator. This fix removes the requirement to specify a security group
when creating a new nodepool through the CLI, otherwise creation fails
when one is not specified.