Switch to use service endpoint for Konnectivity
Using the service endpoint instead of localhost for the konnectivity
server enables kube apiserver rollouts to have access to a working
konnectivity server throughout the rollout. This goes a long way to fix
outages of aggregated APIServices during an APIServer rollout.

This commit also modifies the sync interval for konnectivity agents so
that when a new konnectivity server comes online the amount of time that
it takes for agents to connect to it is reduced.