hypershift-operator: react to hosted cluster changes
In the dedicated request serving node reaper, we want to react to the
deletion of hosted clusters and reap the related nodes. Today, we only
watch nodes and effectively turn our controller into a polling-based
actuator when the hosted clusters still exist. Instead, this change
wires up the controller to watch hosted clusters and enqueue any nodes
related to those clusters. This allows us to remove our polling logic
and react to hosted cluster deletion directly.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>