Validate multi-arch release set & CPU arches match
Validate a multi-arch release image is used when an AWS Hosted Cluster
is set up for multi-arch. Validate management and nodepool CPU arches
match in the CLI if the multi-arch flag is not set. Also, validate each
NodePool CPU arch matches the management cluster's CPU arch when an AWS
Hosted Cluster is not setup for multi-arch.

Signed-off-by: Bryan Cox <brcox@redhat.com>