feat(install): expose hypershift-readers ClusterRole at install time
Today, hypershift install exposes a flag:

```--enable-admin-rbac-generation```
which creates ClusterRoles and ClusterRoleBinding for Hosted cluster admins and readers.

We want to start using the hypershift-readers ClusterRole in ROSA HCP to replace the one currently deployed on the Management Clusters by OSDFM, as it was already agreed that this is better located in HO, as it is close to the resources and the logic.

To do that, we need to:
* add a new install flag
```--enable-reader-rbac-generation```
that creates only the hypershift-readers ClusterRole (as the binding will be handled elsewhere), to avoid creating useless resources that could impact the exposure
* align the ClusterRole with what we have on OSDFM in terms of role aggregation and also add the missing HyperShift API group for certificates which where added recently in HyperShift

The existing flag --enable-admin-rbac-generation is kept unchanged in case we want to deploy both personas including ClusterRoleBindings.

Related: HOSTEDCP-1536