Controller to assist autoscaling for placeholder nodes
Moves warm placeholder pods and sizing placeholder pods to a single
namespace.
Adds a controller that looks at pending placeholder pods, determines a
set of machinesets to scale up to satisfy the placeholders, and scales
up those machinesets. Scaling down is still up to the cluster autoscaler
which will eventually scale down nodes that are not being used.