OCPBUGS-29110: Reconcile over ICSP/IDMS
During the normal reconciliation loop, this patch allow the Hypershift
Operator to grab the new ICSP/IDMS and add them to the CPO, allowing the
HCP pods to use the right internal registry image. The changes on the
ICSP/IDMS will make the control-plane-operator pod to be restarted in
order to update the OPENSHIFT_IMG_OVERRIDES env variable.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>