nodepool_controller: add a reconciler for cleanup
When we change the config for a NodePool or the way in which we hash the
values, we leak token and user data secrets. However, since these
secrets are annotated with the NodePool they were created for, it's
simple to check that the Secret still matches what we expect and clean
it up otherwise.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>