nodepoolcontroller: report performance profile status in `NodePool`
This commit implements status reporting of PerformanceProfile on
Hypershift as captured on:
 https://github.com/openshift/enhancements/pull/1619

Up until now, there was no defined resolution over how PerformanceProfile
status conditions with the status of the different components created
from the PerformanceProfile (MachineConfig, KubeletConfig, Tuned) are handled and exposed.
To resolve this, a custom ConfigMap is added and will be used as a middleware object.
It will hold the updated status populated by the performance profile controller
and will be watched by the NodePool controller, which will mirror the current status provided to the NodePool.status.condition.

The resolution described above consist of the following changes:
* Added a watch to the specific performance profile status ConfigMap object that will trigger the reconcile loop.
* Added the SetPerformanceProfileStatus function to reflect the performance profile status conditions to the nodepool conditions with a prefix.

Co-authored-by: Ronny Baturov <rbaturov@redhat.com>
Signed-off-by: Talor Itzhak <titzhak@redhat.com>