cmd/infra/aws: generate a delegating AWS client
When we run a cleanup task using the myriad credentials that cluster
components have, we need to delegate each call to an AWS service API to
the correct credential that has permissions to use it. We can generate
the delegating client directly from our policy documents to allow us to
keep consumer code from having to know that there are many different
clients operating under the hood in this mode.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>