nto: add configmaps/finalizers under operator `Role`
NTO is watching for the `ConfigMap` that encpasulets the `PerformanceProfile`
and creates different components based on the given profile.

We should block the `PerformanceProfile` deletion as long as the
dependent components exist.

For that NTO needs access the finalizers of the `ConfigMap`.

We should also update the control plane operator since
it's its responsibility to update NTO's role.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>