cmd: DRY out create cluster flag binding
We only bind flags in one routine now; breaking out explicitly the set
of flags that should only be exposed to developers in the `hypershift`
CLI. The net effect of this change is to expose `--base-domain-prefix`
and `--external-dns-domain` to users of `hcp`.

This change also shows how to change the defaults in an option set for a
command - the `hcp create cluster` command has a unique default for the
control plane availability policy, and its now evident that this is the
case since it has to be done explicitly after building the default set
of options.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>