cmd: add an option to render into a file, use it in e2e
The closest thing we have to testing the command-line is our end-to-end
code that re-uses the options structs and runs them through the logic
that the command-line does. We can capture a lot of the work done in the
command-line by inspecting the output manifests, which is easiest to do
by rendering them as a test artfact.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>