Do not ignore single placeholder pods for a specific hosted cluster
The current logic that looks at placeholder pods ignores unpaired pods
because in the case of generic placeholders, the pods are likely in the
middle of getting rolled out by their respective deployment. However,
unpaired pods that correspond to a specific hosted cluster (have a pair
node selector) can exist by themselves if there's already an existing
node for the cluster. This commit updates the logic to not ignore single
placeholder pods for specific hosted clusters.