Allow specifying machine health check timeout in cluster sizing config
Adds an optional machine healthcheck timeout field to the effects of a
size configuration. This allows setting a timeout for specific size of
clusters.

Introduces annotations to override machine health check timeout and max
unavailable values. The annotations can be set in either the
HostedCluster or a specific NodePool. With the one on the NodePool
taking precedence.