Bump openshift/api dep
OpenShift 4.17 has bumped k8s to 1.30, which introduced a change in which ValidatingAdmissionPolicy was promoted to GA and will be enabled by default.
This has caused HCP kube-apiservers to fail because the feature set we were using was setting ValidatingAdmissionPolicy=false, which is now invalid.
Bumping openshift-api dependency should fix the issue by utilizing a valid set of feature gates and to be aligned with ocp 4.17 and k8s 1.30 APIs.

Signed-off-by: Oren Cohen <ocohen@redhat.com>