OCPBUGS-34640: Support for multiple name per component in releaseProvider
This PR adds support for components in the OCP payload that change their names across different OCP releases. Here's how it works: when you invoke the GetImages method, you should provide a slice of strings with the possible names the component could have in different OCP releases. The method will return the first matching name it finds

It also contains a fix for the IgnitionServer covering the component name change for CCO.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>