hypershift:nodepoolcontroller: mirror containerruntime to HCP NS
NTO is one of core OCP operator and it's running as a pod in the HCP namespace.

The operartor requires containerruntime (if exists)
in order to perform certain tuning decisions.

Here, we support mirroring of this object into the HCP namespace,
so the operator will be able to access and get the information it's
needed.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>