OCPBUGS-24400: Recover the ETCD member on cluster boot
Eventually, the ETCD deployment encountered issues booting when many HostedControlPlanes were being created. This caused one or more members to get stuck while trying to join the cluster but unable to do so. Once a pod is stuck in CrashLoopBackOff, the script will now create a snapshot and then remove the member folder, allowing the member to be properly formed and added to the cluster on the next attempt.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>