hack/test: remove timeouts
The top-level `go test` timeout does nothing beneficial for us - we need
to be well-formed in the face of SIGTERM from the test platform anyway,
and we need to time out within our tests, not from the top-level, as
this will give us sensible test failure output (e.g. a specific
assertion failed) versus the current panic output which makes it
impossible to generate jUnit.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>