hack: make the e2e script generic
Today, we duplicate the ci-test-e2e.sh script for Azure, and the
KubeVirt tests don't even use it. This has a bunch of downsides, most
notably that we don't even have jUnit for KubeVirt tests. We should aim
to have only the generic, required, static flags in this script and
allow each caller to simply pass their own as necessary.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>