cmd: use validated and completed opts pattern
Humans make mistakes and the compiler is here to help. The pattern in
this commit is esoteric, but widespread in the Kubernetes ecosystem for
creating options structs that *must* be validated and completed before
being used. The type gymanstics are not ideal but the end experience is
not degraded since we're embedding everything.

In total, the pattern has a couple benefits:
 - enforcement of validation and completion
 - clear distinction between user input (via flags) and computed input
 - obvious flow for re-use in code consumers

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>