Add HTTP(s) konnectivity proxy and use it with OpenShift APIServer
Introduces a new command under cpo: konnectivity-https-proxy that
performs a similar job to konnectivity-socks-proxy except that it
exposes the http protocol. It allows passing user-configured proxy
configuration (http_proxy, https_proxy, no_proxy) as flag values and
uses them when tunneling traffic through konnectivity.

Switches the proxy used by the
OpenShift APIServer to the new proxy and configures it with the proxy
configuration if specified in HostedCluster.spec.configuration.proxy
Fixes volume mounts in the OpenShift APIServer so that additional trust
bundles do not overwrite the default trust bundles in the container.