test/e2e: actually wait for guest cluster client
Creating a new client object does not connect to the API server, so if
we're trying to wait for a client to work, we really need to send an API
request and get a response. The SelfSubjectReview API is simple and we
will always have the privilege to do it, so we can use it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>